requirements.txt
azure-core==1.26.4
 azure-cosmos==4.3.1
 certifi==2022.12.7
 chardet==5.1.0
 charset-normalizer==3.1.0
 click==8.1.3
 colorama==0.4.6
 distlib==0.3.6
 filelock==3.12.0
 idna==3.4
 pip-upgrade==0.0.6
 platformdirs==3.2.0
 pur==7.1.0
 requests==2.28.2
 six==1.16.0
 typing_extensions==4.5.0
 urllib3==1.26.15
 virtualenv==20.22.0
 virtualenv-clone==0.5.7
